feat: integrate db with react ui
