+++ server/db/db.js
@@ -1,0 +1,12 @@
+const Sequelize = require('sequelize')
+const config = require('../../config')
+
+const db = new Sequelize(config);
+
+module.exports = db
+
+// This is a global Mocha hook used for resource cleanup.
+// Otherwise, Mocha v4+ does not exit after tests.
+if (process.env.NODE_ENV === 'test') {
+  after('close database connection', () => db.close())
+}
